move hooks into hooks.ts
