+++ src/popup/components/InputAmount.tsx
@@ -6,8 +6,8 @@ import {
   errorWithKey,
   ErrorWithKeyLike,
   formatCurrency,
-  throttle,
 } from '@/shared/helpers';
+import { useLongPress, useThrottle } from '@/popup/lib/hooks';
 
 interface Props {
   id: string;
@@ -253,42 +253,6 @@ export function validateAmount(
   }
   return null;
 }
-
-function useLongPress(callback: () => void, ms = 100) {
-  const intervalRef = React.useRef<ReturnType<typeof setInterval> | null>(null);
-
-  const start = () => {
-    intervalRef.current ??= setInterval(callback, ms);
-  };
-
-  const stop = () => {
-    if (intervalRef.current) {
-      clearInterval(intervalRef.current);
-      intervalRef.current = null;
-    }
-  };
-
-  React.useEffect(() => stop, [callback, ms]);
-
-  return {
-    onMouseDown: start,
-    onMouseUp: stop,
-    onMouseLeave: stop,
-  };
-}
-
-const useThrottle: typeof throttle = (
-  callback,
-  delay,
-  options = { leading: false, trailing: false },
-) => {
-  const cbRef = React.useRef(callback);
-  React.useEffect(() => {
-    cbRef.current = callback;
-  });
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  return React.useCallback(throttle(cbRef.current, delay, options), [delay]);
-};
 
 function incOrDec(
   input: HTMLInputElement,
